Moved all of the OperatorTranslator classes into 'codegen/operator' #690
