+++ src/codegen/operator/global_group_by_translator.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/global_group_by_translator.h"
+#include "include/codegen/operator/global_group_by_translator.h"
 
 #include "codegen/lang/if.h"
 #include "common/logger.h"
