+++ src/codegen/operator/hash_join_translator.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/hash_join_translator.h"
+#include "include/codegen/operator/hash_join_translator.h"
 
 #include "codegen/lang/if.h"
 #include "codegen/oa_hash_table_proxy.h"
