+++ src/include/codegen/expression/negation_translator.h
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include "codegen/expression_translator.h"
+#include "codegen/expression/expression_translator.h"
 
 namespace peloton {
