+++ src/include/codegen/operator/hash_group_by_translator.h
@@ -16,7 +16,7 @@
 #include "codegen/compilation_context.h"
 #include "codegen/consumer_context.h"
 #include "codegen/oa_hash_table.h"
-#include "codegen/operator_translator.h"
+#include "operator_translator.h"
 #include "codegen/updateable_storage.h"
 
 namespace peloton {
